requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-dotenv==1.1.0
google-generativeai==0.3.1
huggingface-hub==0.32.4
requests==2.32.3
python-multipart==0.0.6
pymongo==4.6.0
motor==3.3.2
groq==0.4.1